Basic implementations of layout functions for tlog-tiles
These functions were copied from the equivalents in serverless-log and then mercilessly hacked down. We may need to restore some of the functions that were removed. We will almost certainly need to add (or re-add, for one function) some computation of the partial tile suffix. That said, getting this in before I finish my work-week feels like a useful foundation on which to build.