Added client for tlog-tiles logs.
This was largely copied from serverless-log and minimally changed to
bind this to the tlog-tiles API which has a few differences from the
serverless API. The obvious changes are the URL paths, but that is
largely hidden behind the `layout` API. The more significant delta is
that serverless returns fully populated tiles, where tlog-tiles returns
only the leaf nodes and thus the client needs to perform some
repopulation of the internal nodes in order to get at all node hashes
for proof generation.

The test data is largely the same as the serverless-log test data in
shape, though it needed to be completely regenerated due to the above
changes. This required the posix log implementation to be ported across
to the tlog-tiles API too. I have this in a local branch and will get it
up in its own PR after this one.