README: direct users to the examples
We can come back and add example usage in this main README if that's useful, but for now pointing to the examples that are fleshed out and are known to compile is a pragmatic step forward. It also gives newcomers links to binaries they can run to understand what Tessera can do, before they start writing any code.