POSIX oneshot: refactor code out of main method (#197)
The goal here is that the main method should contain only the interesting code for putting the entries in the log. This PR moves code into methods to aid readability:

- initializing the verifier
- initializing the signer
- evaluating the glob

This fixes a bug where it would deadlock due to filling entryChan if more than 100 leaves were added at a time.