POSIX conformance: run an HTTP file server (#215)
This simplifies a lot of other code that was spinning up busybox as an http server. This has failed a number of times in CI. The simpler deployment of this approach has a lot going for it. The one downside is that it shows non-canonical usage of the POSIX log. But as there's already an HTTP server running, it seems wasteful to not take advantage in this way.