Wrapper for Add to block until leaf is integrated
The core of this new functionality is a loop that polls for new
checkpoints peridiodically. Client code that calls Await will block
until this polling loop finds a checkpoint that is bigger than the index
being awaited, at which point it releases the client and provides it
with the checkpoint and the index.

This fixes #193.