Clean up integration code (#322)
This code previously had a number of mutexes but no obvious concurrency. Access to some of the mutex-protected fields on populatedTile was questionable; the Equals method had no locks on either object, and the leaves were directly read by another object. Instead of trying to address these by adding more locks, the TreeBuilder object has been moved to be internal, and Integrate is now a function. This makes the state and concurrency much easier to reason about.

Added a benchmark for Integrate.