[README] Docs for personality authors (#342)
Previously the docs for personality authors were essentially "go and
read the example code and figure it out". These docs have now been
expanded to provide an introduction to many useful concepts, and provide
a guide for how to start writing a personality.

This is not intended to be a final revision and it is likely that there
is a lot more content to add here in the coming months.