[MySQL] complete entry bundle implementation (#398)
MySQL API now checks that the number of entries returned in a bundle is not smaller than the number requested.

Now this is fixed, the conformance personality is updated to set cache headers on the entry bundles to allow aggressive caching. Also fixed the checkpoint implementation to never cache, and handle not found properly.

This fixes #364.

While making a change to the SQL schema, took the opportunity to set a couple of other integers to the smallest value needed.