Extract tighter-scoped integrate funcs in storage impls (#440)
This PR pulls the act of updating the merkle tree resources out into separate funcs.

This helps with readability by reducing the size/tightening the scope of some of the funcs in storage implementations (MySQL in particular was getting quite large), but also aids in supporting other lifecycle modes (e.g. for #414).

No functional changes.