[POSIX] Ensure compatibility version (#457)
This will upgrade any existing unversioned logs to have the current version. Towards #450

Regenerated the testdata to confirm this works and checked in the .state
directory, with lock files and all. We may want to remove lock files
from disk when they are unlocked, but that's beyond the scope of this
PR.