[API] Refactoring to make appender API future-proof
The inflexibility of this was noticed while looking at options to fix #341. The draft in this PR shows a Shutdown method, but more interesting are the other changes that a) move the dedupe decorators behind an options, and b) making Appender more than just a function to allow new methods (such as Shutdown) to be added in the future.