Initialise POSIX log if uninitialized without flag
This removes the need to specify an --initialise flag.
If the provided storage_dir does not exist or is empty, the
log will be initialised with the required .state directory.

If the directory contains files, it is assumed to be properly
initialised. I've added a check that the .state directory exists before
creating the required version file, in case a directory is reused that
has other unrelated files.

Fixes #476

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>