[Witnessing] Check responses for valid signatures
This now verifies the body of 200 responses. It checks that the note can
be verified using the signature, and then returns only the signature
that the log has a verifier for.

This means that witnesses that return a valid signature and then a load
of other signatures will not be able to pollute the checkpoint with
these other signatures. On the other hand, it means we will need to
consider how to support witness key rotation in Tessera in the future.
There are a few ways to solve this, but I don't believe this approach
blocks any of them.