[Witness] construct proofs in witness
This switches away from the previous idea which was to pass in a proof
fetcher. Now a tile fetcher is passed in, and the proofs are constructed
in the witness code.

Now that real proofs are being generated, we need to use a real log in
the witness tests. The POSIX log in testdata is used for this purpose.

The unsafe method for parsing checkpoints has also been extended to
return hashes, which was needed to support this work.