Make reads and writes to LST thread safe
Before there was a possibility of a race condition if multiple threads called Update. Such a race had the potential to leave the internal state inconsistent. It could have also allowed clients to see both sides of a split view from the tracker.

Also updated some straggler files that were missed in the field
encapsulation. Whoops.