Removed CheckConsistency method
This method was encouraging people to do the wrong thing and build up a stack of checkpoints before checking them. This is bad; you should only ever have a trusted checkpoint and a potential new candidate for consistency. The method is deleted, and its usages removed and cleaned up. The integration test had a bug before where the array of checkpoints always contained the same checkpoint, so this is also a bug fix.

I'm wrapping up for the week and wanted to get this in. I haven't run this rewritten test yet, so YOLO.