Fix witnessed size state (#510)
Previously the state was being lost. Even though this was unintentional, it worked, but involved constantly querying the witness as if it was at size 0.

This change makes the size sticky for the life of the process by correctly using pointers instead of copying structs. It also updates the size on success, which was missing earlier.