[Client] First pass at tidying the tiles client (#506)
This PR cleans up the client implementation to make it more readable, maintainable, and safer to use.

* Reordered methods

Grouped types, functions, classes, and internal helpers in the more standard order.

* Make LogStateTracker much more private

Only one of its many public properties was being used outside the class. Made a getter for this property, and made all fields private.

* Make reads and writes to LST thread safe

Before there was a possibility of a race condition if multiple threads called Update. Such a race had the potential to leave the internal state inconsistent. It could have also allowed clients to see both sides of a split view from the tracker.

* Removed CheckConsistency method

This method was encouraging people to do the wrong thing and build up a stack of checkpoints before checking them. This is bad; you should only ever have a trusted checkpoint and a potential new candidate for consistency. The method is deleted, and its usages removed and cleaned up. The integration test had a bug before where the array of checkpoints always contained the same checkpoint, so this is also a bug fix.