[Witnessing] Use witnessing in all drivers (#515)
All drivers now take set up witnessing as configured in the AppendOptions. If no options are provided then this is still set up, but does nothing. Towards #309.

This required changing the witness tests package to avoid a package cycle; the posix implementation now uses the witness, but the witness tests use posix.