[Shutdown] Implement Shutdown in a single location
This is a revised attempt at Shutdown that is intended to replace #524.
This implementation doesn't have the granular integration with the
channels in each implementation. Instead, it is implemented at the top
level of the appender lifecycle, and each driver doesn't need to have
very similar duplicated code for handling this.

The previous attempt was also trying to make it so that Shutdown would
close out channels and kill goroutines etc. This approach makes it clear
that the way to do this is to cancel the original context passed into
the appender.

Towards #341.