AWS migration target support
This is based on the GCP migration target implementation.

The conformance cmd was updated to move the listen flag in order to keep
all common AWS configuration flags together, and the flags for the
application logic built on top (e.g. web serving) in another block.

Tested in AWS and it migrated a POSIX log with 189k entries in 7s.