[AWS] Antispam
This adds experimental support for antispam (dedupe) for AWS-hosted
logs. The hashes and indices of seen entries are stored in a MySQL
database.

This is based heavily on the GCP Spanner version, and still maintains
its core way of operation (i.e. filling in the duplicates by using the
Follower pattern).