[AWS Antispam] Fix CI
This is failing at the moment since being enabled in #592. This fixes it
by not creating a second DB and putting the antispam tables in the main
tessera DB. Not elegant, but better than not testing antispam in CI.

This does not change the codelab behaviour; it is still possible to
create the antispam DB from that terraform and run the binaries with the
flags to use it.