[Benchmarks] Log some CPU diagnostics info
We see a fairly clear bimodal performance distribution for our benchmarks. Nothing in the current output indicates any difference between these runs. I have seen other projects include this CPU info in the filenames to compare apples with apples, and oranges with oranges. Before we wire this info into the file name, first dump it to the log so we can confirm that this is different between slower and faster runs.