[API] Extreme refactoring to trial moving core deps out of tessera
Entry, Index, IndexFuture, AddFn, et al, are depended on by almost all code. Having them in the main tessera package meant that extracting anything that depended on them (e.g. Antispam) to an internal package would all but guarantee a package cycle. This PR takes a look at what it would look like if these were moved into a separate package.

In my mind, this is similar to a proto package that you might find in a standard gRPC app. These core constructs are really intended to be types that everything else uses, and that should depend on very little.